Add CI build with docker capabilities
The build has been failing due to needing docker for SampleTestRunnerTests and it not being available on the docker-based executors (CircleCI's docker executors do not allow docker-in-docker). A machine executor was already defined in config.yml but a build that used it was not. This adds such a build and updates the build configuration to have a task (`dockerTest`) that runs only tests requiring docker (determined by having a `@Tag` with value `docker` on them).