Tracing handler is not setting on the Observation context a previous scope
without this change we're overriding the scopes in observation context. Since observation context is mutable it can be executed multiple times by the same handler (e.g. you open a scope multiple times for the same observation). That means that the created tracing scope must stack the previous one, so that when close is called we set the previous scope as current in the observation context.

with this change we modify the open scope logic so that when a new tracing scope is opened we store it on the observation context but on closing we restore the previous scope that was there (or null when there was no previous one)

fixes gh-150