TracingContext contains now scope per thread
without this change TracingContext contains one reference to multiple scopes. Brave requires a concrete instance of the scope to be present upon closing the scope, otherwise it will throw an assertion error.

with this change we're ensuring that we store a scope per thread. That way we don't make one thread override another ones scope

fixes gh-157