Annotation parity with Sleuth
adds missing implementations of interfaces for @NewSpan and @ContinueSpan aspects.

feature parity will not be full cause we don't have access to Spring code in Micrometer so aspects will not work as well as they did in Sleuth. The difference will be that annotations (@NewSpan and @ContinueSpan) on interfaces will not be supported. We will only support them on the implementing classes, however we will support @SpanTag on the interfaces too).

fixes gh-163