Not creating scopes when resetting scope on trace handler
without this change, we're using the `tracer.withSpan(null)` to reset the thread local, however the better alternative is to use the `CurrentTraceContext.maybeScope` mechanism.

with this change we're changing the way the `TracingObservationHandler` is clearing scopes. We also needed to improve the testing tracer to support these changes

fixes gh-213