Fix scoping
without this change we are not properly resetting scopes. When onReset is called we can still leave some spans behind.

with this change we additionally call brave's and Otel mechanism to ensure that no span is left in thread local. Also, we add a lot of new tests including reactor ones.