OtelSpanBuilder now uses SpanBuilder properly
without this change we would first create a SpanBuilder with an empty name and then update the name on the span. There are cases that depending on the name of the span OTel can already interact with the SpanBuilder and alter its behaviour.

with this change SpanBuilder is created on the start() method where it should already contain all the necessary information including the proper name. We set the name on the builder instead of the span.

fixes gh-259