Doesn't clear scopes
without this change we're clearing the scopes map upon ending of spans. The problem is that the scoping hierarchy gets broken and even though that OTLA works well (scopes are properly opened and closed) the spans are not.

with this change we're not removing the scopes upon clearing of a TraceContext

fixds gh-363