Fixed wrong OtelTraceContext sampling and parent fields resolution
Without this change an NPE was thrown when unboxing the sampling flag. Also the parent id was resolved improperly. Lastly OTel doesn't understand a deferred sampling decision (that's why NPE was thrown, but even after fixing it sampling decision is either positive or negative).
with this change we're fixing the unboxing problem plus we're directly passing the sampling decision and parent id as respones for trace context method execution (for sampling and parent id)

fixes gh-911