Remove default span name transformation
Previous behavior would transform the name taken from the Observation Context by making it kebab case. It's not clear this behavior was ever intentionally applied to all spans created from the DefaultTracingObservationHandler. Nonetheless, users may have come to rely on this transformation, and for them, this is a breaking change. We still have time to revisit this before the Release Candidate.